Remove hashes from PHP code, condense
The PHP file is made specifically for being called through AJAX, and isn't very useful for programmers to link to in their own PHP files (the getFiles function could be the exception). I've removed the class here and made it a simpler file, to try to make things more straightforward and add a touch of performance.

Hashes were removed because in this context, it really didn't make sense. I'd need to create a whole account system for users at that point to allow hashes, and possible MySQL requirements, etc; I couldn't think of a good way to use hashes that overall INCREASED security without likely adding more holes. Plain-text password in the PHP file isn't great, but I think in this case trying to work with that would just add even MORE security vulnerabilities. If users just use a unique password, then if they're hacked they'll be fine.