Fix bugs, improve Hey Bard support, add time-based bookmarks
Bookmarks and querystrings used to only be based on file number. Now they can be based on time instead- great for video series and the like.

Also changed the name of S.object to S.HeyBardID to more clearly get across its purpose, and added the S.bookmark value to specify whether to save bookmarks and querystrings as file numbers of timestamps (in seconds).